+++ tests/test_tutorial/test_sql_databases/test_sql_databases_middleware.py
@@ -1,3 +1,4 @@
+import importlib
 from pathlib import Path
 
 import pytest
@@ -283,13 +284,19 @@ openapi_schema = {
 
 @pytest.fixture(scope="module")
 def client():
-    # Import while creating the client to create the DB after starting the test session
-    from sql_databases.sql_app.alt_main import app
-
     test_db = Path("./sql_app.db")
-    with TestClient(app) as c:
+    if test_db.is_file():  # pragma: nocover
+        test_db.unlink()
+    # Import while creating the client to create the DB after starting the test session
+    from sql_databases.sql_app import alt_main
+
+    # Ensure import side effects are re-executed
+    importlib.reload(alt_main)
+
+    with TestClient(alt_main.app) as c:
         yield c
-    test_db.unlink()
+    if test_db.is_file():  # pragma: nocover
+        test_db.unlink()
 
 
 def test_openapi_schema(client):
